requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
fastapi
uvicorn
bottle
yfinance
requests
readline

# utilities
six>=1.14
joblib
memoized-property
termcolor
http3
yahoo_fin
